fix install third rom

+++ src/com/grarak/romswitcher/Utils/ChooseRom.java
@@ -38,8 +38,12 @@ public class ChooseRom {
 	private static int buffKey = 0;
 	private static CharSequence[] choiceList;
 
-	private static String SECOND_ROM = "/data/media/.secondrom/system.img";
-	private static String THIRD_ROM = "/data/media/.thirdrom/system.img";
+	private static String DATA;
+	private static final String SECOND_ROM = DATA
+			+ "/media/.secondrom/system.img";
+	private static final String THIRD_ROM = DATA
+			+ "/media/.thirdrom/system.img";
+	private static final String ROM_SELECTION = DATA + "/media/.rom";
 
 	private static final File mSecondSystem = new File(SECOND_ROM);
 	private static final File mThirdSystem = new File(THIRD_ROM);
@@ -49,6 +53,12 @@ public class ChooseRom {
 			final String firstrom, final String secondrom, final String thirdrom) {
 		Builder builder = new Builder(context);
 		builder.setTitle(title);
+
+		if (mSecondRom.exists()) {
+			DATA = "/.firstrom";
+		} else {
+			DATA = "/data";
+		}
 
 		List<String> listItems = new ArrayList<String>();
 		listItems.add(firstrom);
@@ -88,15 +98,15 @@ public class ChooseRom {
 										.equals(secondrom)) {
 									flashkernel("second", context);
 									selected = buffKey;
-									Utils.runCommand(
-											"echo 1 > /data/media/.rom", 0);
+									Utils.runCommand("echo 1 > "
+											+ ROM_SELECTION, 0);
 									((Activity) context).finish();
 								} else if (choiceList[buffKey].toString()
 										.equals(thirdrom)) {
 									flashkernel("second", context);
 									selected = buffKey;
-									Utils.runCommand(
-											"echo 2 > /data/media/.rom", 0);
+									Utils.runCommand("echo 2 > "
+											+ ROM_SELECTION, 0);
 									((Activity) context).finish();
 								}
 							}
@@ -117,11 +127,13 @@ public class ChooseRom {
 	}
 
 	private static void flashkernel(String rom, Context context) {
-		Utils.runCommand("dd if=" + getExternalStorageDirectory().getPath()
-				+ "/romswitcher/" + rom + ".img of="
-				+ SupportedDevices.bootpartition
-				+ " && echo 1 > /proc/sys/kernel/sysrq"
-				+ " && echo b > /proc/sysrq-trigger", 0);
+		if (!mSecondRom.exists() && !rom.equals("first")) {
+			Utils.runCommand("dd if=" + getExternalStorageDirectory().getPath()
+					+ "/romswitcher/" + rom + ".img of="
+					+ SupportedDevices.bootpartition
+					+ " && echo 1 > /proc/sys/kernel/sysrq"
+					+ " && echo b > /proc/sysrq-trigger", 0);
+		}
 		Utils.runCommand("reboot", 1);
 	}
 }
